requirements_hf.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
openai==1.3.7
google-generativeai==0.4.0
sentence-transformers==2.2.2
gradio==4.19.1
torch==2.1.0
transformers==4.35.0
accelerate==0.24.0